Factor out the delta computation
